74 create test dataset

+++ tests/conftest.py
@@ -1,6 +1,8 @@
 import os
 import pytest
 
+from glob import glob
+
 TEST_CONFIG = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'cosmoconfig_test.yaml')
 
 # Check to make sure that the test config file is being used. If not, don't run the tests
@@ -13,7 +15,8 @@ def db_cleanup():
     yield  # The tests don't actually need this test "value"
 
     # Cleanup
-    os.remove('test.db')   # Delete test database file after the completion of all tests
+    if os.path.exists('test.db'):
+        os.remove('test.db')   # Delete test database file after the completion of all tests
 
     # Remove temporary shared memory file if it exists
     if os.path.exists('test.db-shm'):
@@ -22,3 +25,24 @@ def db_cleanup():
     # Remove temporary write-ahead log file if it exists
     if os.path.exists('test.db-wal'):
         os.remove('test.db-wal')
+
+
+@pytest.fixture(scope='session')
+def data_dir():
+    return os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data/')
+
+
+@pytest.fixture(scope='session')
+def here():
+    return os.path.dirname(os.path.abspath(__file__))
+
+
+@pytest.fixture(scope='session', autouse=True)
+def clean_up_output(here):
+    yield
+
+    output = glob(os.path.join(here, '*html')) + glob(os.path.join(here, '*csv'))
+
+    if output:
+        for file in output:
+            os.remove(file)
